+++ lib/delayed/heartbeat.rb
@@ -34,11 +34,11 @@ module Delayed
 
       def cleanup_workers(workers, mark_attempt_failed: true)
         Delayed::Heartbeat::Worker.transaction do
-          orphaned_jobs = workers.flat_map do |worker|
-            worker.unlock_jobs(mark_attempt_failed: mark_attempt_failed)
+          worker_job_map = workers.each_with_object(Hash.new) do |worker, worker_job_map|
+            worker_job_map[worker] = worker.unlock_jobs(mark_attempt_failed: mark_attempt_failed)
           end
           Delayed::Heartbeat::Worker.delete_workers(workers)
-          Delayed::Heartbeat::DeleteWorkerResults.new(workers, orphaned_jobs)
+          Delayed::Heartbeat::DeleteWorkerResults.new(worker_job_map)
         end
       end
     end