+++ spec/delayed/heartbeat/delete_worker_results_spec.rb
@@ -3,32 +3,29 @@ require 'spec_helper'
 describe Delayed::Heartbeat::DeleteWorkerResults do
   let(:worker) { create_worker(name: 'my-worker') }
   let(:job) { create_job(locked_by: worker.name) }
-  let(:results) { create_results([worker], [job]) }
+  let(:results) { create_results(worker => [job]) }
 
   describe "#workers" do
     specify { expect(results.workers).to eq [worker] }
   end
 
   describe "#unlocked_jobs" do
-    specify { expect(results.unlocked_jobs).to eq [job] }
+    let(:other_worker) { create_worker(name: 'my-other-worker') }
+    let(:other_job) { create_job(locked_by: worker.name) }
+    let(:results) { create_results(worker => [job], other_worker => [other_job]) }
+
+    specify { expect(results.unlocked_jobs(worker)).to eq [job] }
+    specify { expect(results.unlocked_jobs).to match_array [job, other_job] }
   end
 
   describe "#to_s" do
     it "includes workers" do
-      results = create_results([worker], [])
+      results = create_results(worker => [])
       expect(results.to_s).to include(worker.name)
     end
 
-    it "includes jobs for known workers" do
-      results = create_results([worker], [job])
+    it "includes jobs for workers" do
       expect(results.to_s).to include(job.id.to_s)
-    end
-
-    it "includes jobs locked by an unknown worker" do
-      job = create_job(locked_by: 'unknown-worker')
-      results = create_results([], [job])
-      expect(results.to_s).to include(job.id.to_s)
-      expect(results.to_s).to include(job.locked_by)
     end
   end
 
@@ -41,7 +38,7 @@ describe Delayed::Heartbeat::DeleteWorkerResults do
     Delayed::Job.create!(attributes)
   end
 
-  def create_results(workers, unlocked_jobs)
-    Delayed::Heartbeat::DeleteWorkerResults.new(workers, unlocked_jobs)
+  def create_results(worker_job_map)
+    Delayed::Heartbeat::DeleteWorkerResults.new(worker_job_map)
   end
 end