+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
 ### 0.2.0 (unreleased)
-* 
+* Enable correlation between workers and unlocked jobs in results returned from deleting workers